Fixed bug where QuarterFilter page wasn't showing up, moved page under Admin dropdown
